Add files via upload
the model is pretty accurate, but for the medical field this is not enough. Not diagnosing a tumor correctly can lead to someones death so this will not be acceptable. In the medical field an accuracy of 99,9% is acceptable.
Next to that you rather have a false positive than a false negative. With a false negative treatment of the Malignant tumor will not begin and it is possible that this will kill the person